move toolbox.js to panels folder
